Fix issues with retaining insertion order in collections (#5)
* Convert many uses of Collectors to CollectorUtils

The collections made through CollectorUtils are guaranteed to retain
insertion order.

There are two specific uses of Collectors#toMap in MergeMetadata which
were not changed to CollectorUtils, as insertion order does not
matter for those collections due to only being used as lookups.

* Switch call to findOverrides to remove TreeSet usage

Using the existing override, which supplies the insertion-respecting
LinkedHashSet, is better.

* Switch to LinkedHash* collections

Every place where a collection needs to retain insertion order -- such
as because its contents are iterated over or stored in the final
metadata -- has been changed to use a LinkedHash map or set.

* Update Feather to 1.1.0

This includes the fix to the metadata classes to use insertion order retaining collections.