Uplift map and ASMRemapper creation outside of class processing (#6)
Previously, the map of obfuscated to Mojang method names and the
obfuscated to Mojang ASMRemapper were recreated each time a class was
processed, which is wasteful as nothing changes between each instance.

This moves the creation of both up a step, outside the individual class
processing, and passes the two to the methods which previously used
them. This results in a massive speedup of the MergeMetadata task, with
about a 10x speedup.